split start into start_server and start_client
