+++ src/common/errors.rs
@@ -1,9 +1,8 @@
 //! Error context system based on a thread-local representation of the call stack, itself based on
 //! the instructions that are sent between threads.
 
-use super::{
-    os_input_output::ServerOsApiInstruction, AppInstruction, ServerInstruction, OPENCALLS,
-};
+use super::{os_input_output::ServerOsApiInstruction, ServerInstruction, OPENCALLS};
+use crate::client::AppInstruction;
 use crate::pty_bus::PtyInstruction;
 use crate::screen::ScreenInstruction;
 use serde::{Deserialize, Serialize};
